Rewrite some code.
Remove method createBoard in Checkers class. This function is called in Man::queenMove so it was moved.
Remove method moveFigure in Board class and create function moveFigure in Man.h file. After that moveFigure function does what its name says and does it as it is supposed to be.

Add methods of Man class: queenMove, onQueenPosition and onQueenPositionIf

Handle the situation, when a man (pawn) becomes a queen (king) killing an opponent.